Update components and added eslint config
