+++ server/.env.prod
@@ -1,0 +1,8 @@
+DB_DIALECT=postgres
+DB_HOST=roteirizador.postgres.database.azure.com
+DB_USER=catel
+DB_PASS=Msc@2014
+DB_NAME=roteirizador
+DB_STORAGE=src/database/database.sqlite
+DB_PORT=5432
+DB_SSL=true
