+++ web/src/App.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Routes from './routes';
 
-import './App.css';
+import './assets/styles/App.css';
 
 function App() {
   return (
