+++ web/src/assets/styles/App.css
