--- web/src/components/MapContainer/styles.css
@@ -1,22 +1,0 @@
-#map {
-    width: 100%;
-    height: 30rem;
-    position: relative;
-    min-width: 25rem;
-}
-
-.loading {
-    display: flex;
-    width: 100%;
-    height: 100%;
-    justify-content: center;
-    align-items: center;
-}
-
-@media(min-width: 1000px) {
-    #map {
-        width: 100%;
-        height: 55rem;
-        position: relative;
-    }
-}
