+++ web/src/services/api.js
@@ -1,7 +1,7 @@
 import axios from 'axios';
 
 const api = axios.create({
-    baseURL: 'http://localhost:3333'
+  baseURL: process.env.REACT_APP_BACKEND_URL || 'http://localhost:3333',
 });
 
 export default api;
