+++ web/src/services/saveRoute.js
@@ -1,22 +1,18 @@
 import api from './api';
 
-async function saveRoute(coordinates) {
-
-    if (typeof coordinates !== 'undefined' && coordinates.length > 0) {
-
-        await api.post("routes", {
-            coordinates: coordinates.map(coordinate => ({
-                longitude: coordinate.lat,
-                latitude: coordinate.lng
-            }))
-        }).then(response => {
-            return response;
-        }).catch(err => {
-            console.error(err);
-            alert("Erro ao salvar a rota no banco de dados");
-            return null;
-        });
+const saveRoute = async (coordinates) => {
+  try {
+    if (coordinates && coordinates.length > 0) {
+      await api.post('routes', {
+        coordinates: coordinates.map((coordinate) => ({
+          longitude: coordinate.lat,
+          latitude: coordinate.lng,
+        })),
+      });
     }
-}
+  } catch (error) {
+    console.log(error);
+  }
+};
 
 export default saveRoute;
