--- web/src/utils/constants.js
@@ -1,3 +1,0 @@
-const apiKey = "AIzaSyC09lOsJr2ZebDd2CHt_3X0JCoP-yrZUJk";
-
-export default apiKey;
